requirements.txt
asgiref==3.5.2
attrs==22.1.0
awesomeversion==22.9.0
beautifulsoup4==4.11.1
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.7
django-allauth==0.51.0
django-cors-headers==3.11.0
django-extensions==3.1.5
django-filter==21.1
django-js-asset==2.0.0
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.4
flake8==6.0.0
gunicorn==20.1.0
idna==3.4
importlib-resources==5.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.1
MarkupSafe==2.1.1
mccabe==0.7.0
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==21.3
pkgutil_resolve_name==1.3.10
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.0
pyimgur==0.6.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
swagger-spec-validator==3.0.3
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
wincertstore==0.2
zipp==3.10.0
